refactor: migrate mapper into module from file
