+++ argilla/src/argilla/records/_mapping/_mapper.py
@@ -14,11 +14,8 @@
 
 import re
 import warnings
-from enum import Enum
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Sequence, Union, Tuple
 from uuid import UUID
-
-from pydantic import BaseModel, Field
 
 from argilla.records._resource import Record
 from argilla.responses import Response
@@ -26,111 +23,32 @@ from argilla.settings import TextField, VectorField
 from argilla.settings._metadata import MetadataPropertyBase
 from argilla.settings._question import QuestionPropertyBase
 from argilla.suggestions import Suggestion
+from argilla.records._mapping._routes import (
+    AttributeRoute,
+    RecordAttributesMap,
+    AttributeType,
+    ParameterType,
+    AttributeParameter,
+)
 
 if TYPE_CHECKING:
     from argilla.datasets import Dataset
 
 
-class ParameterType(str, Enum):
-    """Parameter types are the different 'sub' values of a records attribute.
-    For example, the value, score, or agent of a suggestion."""
-
-    VALUE = "value"
-    SCORE = "score"
-    AGENT = "agent"
-
-    @classmethod
-    def values(cls) -> List[str]:
-        return [param.value for param in cls]
-
-
-class AttributeType(str, Enum):
-    """Attribute types are the different types of attributes a record can have."""
-
-    FIELD = "field"
-    SUGGESTION = "suggestion"
-    RESPONSE = "response"
-    METADATA = "metadata"
-    VECTOR = "vector"
-    ID = "id"
-
-    @classmethod
-    def values(cls) -> List[str]:
-        return [attr.value for attr in cls]
-
-
-class AttributeParameter(BaseModel):
-    """Attribute parameters are the different 'sub' values of a records attribute.
-    And the source in the data that the parameter is coming from.
-    """
-
-    parameter_type: ParameterType = ParameterType.VALUE
-    source: str
-
-
-class AttributeRoute(BaseModel):
-    """AttributeRoute is a representation of a record attribute that is mapped to a source value in the data."""
-
-    source: str
-    name: str
-    type: Optional[AttributeType] = None
-    parameters: List[AttributeParameter] = []
-
-    def set_parameter(self, parameter: AttributeParameter):
-        """Set a parameter for the route.
-        An existing parameter with same parameter type will be replaced by this new one.
-        """
-        for p in self.parameters:
-            if p.parameter_type == parameter.parameter_type:
-                self.parameters.remove(p)
-                break
-        self.parameters.append(parameter)
-
-
-class RecordAttributesMap(BaseModel):
-    """RecordAttributesMap is a representation of a record attribute mapping that is used to parse data into a record."""
-
-    suggestion: Dict[str, AttributeRoute] = Field(default_factory=dict)
-    response: Dict[str, AttributeRoute] = Field(default_factory=dict)
-    field: Dict[str, AttributeRoute] = Field(default_factory=dict)
-    metadata: Dict[str, AttributeRoute] = Field(default_factory=dict)
-    vector: Dict[str, AttributeRoute] = Field(default_factory=dict)
-
-    id: AttributeRoute = AttributeRoute(source="id", name="id", type=AttributeType.ID)
-
-    def _get_routes_group_by_type(self, type: AttributeType):
-        return {
-            AttributeType.SUGGESTION: self.suggestion,
-            AttributeType.RESPONSE: self.response,
-            AttributeType.FIELD: self.field,
-            AttributeType.METADATA: self.metadata,
-            AttributeType.VECTOR: self.vector,
-            AttributeType.ID: self.id,
-        }[type]
-
-    def get_by_name_and_type(self, name: str, type: AttributeType) -> Optional[AttributeRoute]:
-        """Get a route by name and type"""
-        if name == "id" and AttributeType.ID:
-            return self.id
-        return self._get_routes_group_by_type(type).get(name)
-
-    def add_route(self, attribute_route: AttributeRoute) -> None:
-        """Ad a new mapping route"""
-        if attribute_route.type == AttributeType.ID:
-            self.id = attribute_route
-        else:
-            self._get_routes_group_by_type(attribute_route.type)[attribute_route.name] = attribute_route
-
-
 class IngestedRecordMapper:
     """IngestedRecordMapper is a class that is used to map data into a record object.
-    It maps values in ingested data to the appropriate record attributes, based on the user provided mapping and the schema of the dataset.
+    It maps values in ingested data to the appropriate record attributes, based on the user
+    provided mapping and the schema of the dataset.
+
+    The Mapper builds and uses a `RecordAttributesMap` object to map the data to the appropriate record attributes.
 
     Attributes:
         dataset: The dataset the record will be added to.
         mapping: A dictionary mapping from source data keys/ columns to Argilla fields, questions, ids, etc.
         user_id: The user id to associate with the record responses.
     """
+
+    mapping: RecordAttributesMap = None
 
     def __init__(
         self,
@@ -187,30 +105,30 @@ class IngestedRecordMapper:
 
         for source_key, value in mapping.items():
             mapped_attributes = [value] if isinstance(value, str) else list(value)
-            for attribute_mapping in mapped_attributes:
-                attribute_name, attr_type, parameter = self._parse_dot_notation(attribute_mapping)
+            for attr_mapping in mapped_attributes:
+                attr_name, attr_type, parameter = self._parse_dot_notation(attr_mapping)
 
                 attr_type = AttributeType(attr_type or AttributeType.SUGGESTION)
                 parameter = AttributeParameter(parameter_type=parameter or ParameterType.VALUE, source=source_key)
 
-                attribute_route = default_mapping.get_by_name_and_type(name=attribute_name, type=attr_type)
-                if attribute_route:
-                    attribute_route.source = source_key
-                    attribute_route.set_parameter(parameter)
+                attr_route = default_mapping.get_by_name_and_type(name=attr_name, type=attr_type)
+                if attr_route:
+                    attr_route.source = source_key
+                    attr_route.set_parameter(parameter)
                 else:
-                    attribute_route = AttributeRoute(
-                        name=attribute_name,
+                    attr_route = AttributeRoute(
+                        name=attr_name,
                         source=source_key,
                         type=attr_type,
                         parameters=[parameter],
                     )
-                    attribute_route = self._select_attribute_type(attribute=attribute_route)
-                    default_mapping.add_route(attribute_route)
+                    attr_route = self._select_attribute_type(attribute_route=attr_route)
+                    default_mapping.add_route(attr_route)
 
         return default_mapping
 
     def _parse_dot_notation(self, attribute_mapping: str) -> Tuple[str, Optional[str], Optional[str]]:
-        """Parses a string in the format of 'attribute.type.parameter' into its parts using regex."""
+        """Parses a string in the format of 'attribute.type.parameter' into its attribute parts parts using regex."""
 
         available_attributes = list(self._schema.keys()) + ["id"]
         available_parameters = ParameterType.values()
@@ -230,44 +148,44 @@ class IngestedRecordMapper:
                 "Attribute mapping must be in the format of 'attribute[.type[.parameter]]'."
                 f"Available attributes: {available_attributes}, types: {available_types}, parameters: {available_parameters}."
             )
-        attribute_name, type_, parameter = match.groups()
-        return attribute_name, type_, parameter
+        attr_name, type_, parameter = match.groups()
+        return attr_name, type_, parameter
 
-    def _select_attribute_type(self, attribute: AttributeRoute) -> AttributeRoute:
+    def _select_attribute_type(self, attribute_route: AttributeRoute) -> AttributeRoute:
         """Selects the attribute type based on the schema item and the attribute type.
         This method implements the logic to infer the attribute type based on the schema item if the attribute type is not provided.
         If the attribute type is not provided, it will be inferred based on the schema item.
         """
-        schema_item = self._schema.get(attribute.name)
+        schema_item = self._schema.get(attribute_route.name)
         if isinstance(schema_item, QuestionPropertyBase) and (
-            attribute.type is None or attribute.type == AttributeType.SUGGESTION
+            attribute_route.type is None or attribute_route.type == AttributeType.SUGGESTION
         ):
             # Suggestions are the default destination for questions.
-            attribute.type = AttributeType.SUGGESTION
-        elif isinstance(schema_item, QuestionPropertyBase) and attribute.type == AttributeType.RESPONSE:
-            attribute.type = AttributeType.RESPONSE
+            attribute_route.type = AttributeType.SUGGESTION
+        elif isinstance(schema_item, QuestionPropertyBase) and attribute_route.type == AttributeType.RESPONSE:
+            attribute_route.type = AttributeType.RESPONSE
         elif isinstance(schema_item, TextField):
-            attribute.type = AttributeType.FIELD
+            attribute_route.type = AttributeType.FIELD
         elif isinstance(schema_item, VectorField):
-            attribute.type = AttributeType.VECTOR
+            attribute_route.type = AttributeType.VECTOR
         elif isinstance(schema_item, MetadataPropertyBase):
-            attribute.type = AttributeType.METADATA
-        elif attribute.name == "id":
-            attribute.type = AttributeType.ID
+            attribute_route.type = AttributeType.METADATA
+        elif attribute_route.name == "id":
+            attribute_route.type = AttributeType.ID
         else:
-            warnings.warn(message=f"Record attribute {attribute.name} is not in the schema or mapping so skipping.")
-        return attribute
+            warnings.warn(message=f"Record attribute {attribute_route.name} is not in the schema or mapping so skipping.")
+        return attribute_route
 
     def _schematize_default_attributes(self) -> RecordAttributesMap:
-        """Creates the mapping with default attributes. Uses the schema of the dataset to infer
-         the default attributes and add them to the mapping.
+        """Creates the mapping with default attribute routes. Uses the schema of the dataset to determine
+         the default attributes and add them to the mapping with their names as keys. This means that 
+         keys in the data that match the names of dataset attributes will be mapped to them by default.
 
         Returns:
             RecordAttributesMap: The mapping object.
         """
         mapping = RecordAttributesMap()
 
-        # Map keys that match question names to the suggestion attribute type.
         for question in self._dataset.settings.questions:
             mapping.suggestion[question.name] = AttributeRoute(
                 source=question.name,
@@ -321,6 +239,8 @@ class IngestedRecordMapper:
             if route.source not in data:
                 continue
             parameters = {param.parameter_type: data.get(param.source) for param in route.parameters}
+            if parameters.get(ParameterType.VALUE) is None:
+                continue
             schema_item = self._dataset.schema.get(name)
             suggestion = Suggestion(
                 **parameters,
@@ -345,8 +265,11 @@ class IngestedRecordMapper:
         responses = []
 
         for name, route in mapping.items():
+            value = data.get(route.source)
+            if value is None:
+                continue
             response = Response(
-                value=data.get(route.source),
+                value=value,
                 question_name=name,
                 user_id=user_id,
             )
@@ -357,10 +280,13 @@ class IngestedRecordMapper:
     def _map_attributes(self, data: Dict[str, Any], mapping: Dict[str, AttributeRoute]) -> Dict[str, Any]:
         """Converts a dictionary to a dictionary of attributes for use by the add or update methods."""
         attributes = {}
+        
         for name, route in mapping.items():
             if route.source not in data:
                 continue
             value = data.get(route.source)
-            if value is not None:
-                attributes[name] = value
+            if value is None:
+                continue
+            attributes[name] = value
+            
         return attributes
